put each component into a separate file
