+++ lib/GridCell.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default (props) => <td>{props.children}</td>;
