+++ lib/GridTable.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import Table from 'react-bootstrap/lib/Table'
+
+export default (props) => <Table bordered hover>{props.children}</Table>;
