+++ lib/constants.js
@@ -1,0 +1,2 @@
+export const ASCENDING = "asc";
+export const DESCENDING = "desc";
